# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/meal-plans": {
      "post": {
        "operationId": "MealPlanController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The meal plan has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlanBaseResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new meal plan",
        "tags": [
          "meal-plans"
        ]
      },
      "get": {
        "operationId": "MealPlanController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all meal plans.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlansBaseResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all meal plans",
        "tags": [
          "meal-plans"
        ]
      }
    },
    "/meal-plans/user": {
      "get": {
        "operationId": "MealPlanController_findAllByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return user meal plans.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlansBaseResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user meal plans",
        "tags": [
          "meal-plans"
        ]
      }
    },
    "/meal-plans/{id}": {
      "get": {
        "operationId": "MealPlanController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meal plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the meal plan with specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlanBaseResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Meal plan not found."
          }
        },
        "summary": "Get a meal plan by ID",
        "tags": [
          "meal-plans"
        ]
      },
      "patch": {
        "operationId": "MealPlanController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meal plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The meal plan has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlanBaseResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Meal plan not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a meal plan",
        "tags": [
          "meal-plans"
        ]
      },
      "delete": {
        "operationId": "MealPlanController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meal plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The meal plan has been successfully deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Meal plan not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a meal plan",
        "tags": [
          "meal-plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Meal Plan API",
    "description": "Meal Plan API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateMealPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the meal plan",
            "example": "Weight Loss Plan"
          },
          "durationInDays": {
            "type": "number",
            "description": "Duration of the meal plan in days",
            "example": 30
          },
          "mealCharacteristicId": {
            "type": "string",
            "description": "Meal characteristic UUID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "description": "User UUID",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        },
        "required": [
          "name",
          "durationInDays"
        ]
      },
      "MealPlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Meal plan UUID",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "name": {
            "type": "string",
            "description": "Name of the meal plan",
            "example": "Weight Loss Plan"
          },
          "durationInDays": {
            "type": "number",
            "description": "Duration of the meal plan in days",
            "example": 30
          },
          "mealCharacteristicId": {
            "type": "string",
            "description": "Meal characteristic UUID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "description": "User UUID",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2023-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2023-01-02T00:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "durationInDays",
          "createdAt",
          "updatedAt"
        ]
      },
      "MealPlanBaseResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Meal plan data",
            "allOf": [
              {
                "$ref": "#/components/schemas/MealPlanResponseDto"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "MealPlansBaseResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of meal plans",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealPlanResponseDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateMealPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Meal plan UUID",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "name": {
            "type": "string",
            "description": "Name of the meal plan",
            "example": "Modified Weight Loss Plan"
          },
          "durationInDays": {
            "type": "number",
            "description": "Duration of the meal plan in days",
            "example": 45
          },
          "mealCharacteristicId": {
            "type": "string",
            "description": "Meal characteristic UUID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}